Actualizacion Service - Interface - DTOs - Entidades

+++ src/main/java/com/bitraid/tuktuktaxi/repositories/DriverRepository.java
@@ -3,5 +3,5 @@ package com.bitraid.tuktuktaxi.repositories;
 import com.bitraid.tuktuktaxi.entities.Driver;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DriverRepository extends JpaRepository<Driver, Integer> {
+public interface DriverRepository extends JpaRepository<Driver, Long> {
 }
