+++ src/main/java/com/bitraid/tuktuktaxi/repositories/PassengerRepository.java
@@ -3,5 +3,5 @@ package com.bitraid.tuktuktaxi.repositories;
 import com.bitraid.tuktuktaxi.entities.Passenger;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PassengerRepository extends JpaRepository<Passenger, Integer> {
+public interface PassengerRepository extends JpaRepository<Passenger, Long> {
 }
